feat(mixin): add ConditionalAuditRepositoryMixin for configurable Audit Mixin
To be used when the actual audit mixin is to be configured from the flag in the environment variable
rather than the code

BREAKING CHANGE:
Audit Repository Mixin return type changed and now is being returned as an abstract class. This
doesn't affect 99% of the users because this mixin is already applied on the base class. Only those
who are directly assigned the mixin-ed class to some variable will have the type error showing them.

GH-57