feat(mixin): add conditional audit repository mixin (#58)
To be used when the actual audit mixin is to be applied conditionally from the flag from the environment variables
rather than the code.

BREAKING CHANGE:
Audit Repository Mixin return type changed and now is being returned as an abstract class. This
doesn't affect 99% of the users because in crud repositories this mixin is already applied on the base class. Only those who are directly assigning the mixin-ed class to some variable will have the type error showing them.

GH-57